+++ oidc-controller/src/VCAuthn/Utils/ProofRequestUtils.cs
@@ -37,15 +37,5 @@ namespace VCAuthn.Utils
         }
     }
 
-    class ProofRequest_v_1_0
-    {
-        [JsonProperty("version")]
-        public string Version { get; set; }
-
-        [JsonProperty("requested_attributes")]
-        public Dictionary<string, RequestedAttribute> RequestedAttributes { get; set; } = new Dictionary<string, RequestedAttribute>();
-
-        [JsonProperty("requested_predicates")]
-        public Dictionary<string, RequestedPredicate> RequestedPredicates { get; set; } = new Dictionary<string, RequestedPredicate>();
-    }
+    
 }
